requirements.txt
viam-sdk
opencv-python
numpy 